Fix validation errors in the RSS and Atom Feeds (#761)
## Proposed changes

The encoded html in the content of atom and RSS feeds can not have
relative urls in them. Use the transformTag method of the sanitize-html
plugin to switch anchor tag href and image src properties to be
absolute. Load all the images up ahead of time and store them in a map
with a normalized key so that we can later look them up and get the
absolute url for them. Fixes #760

## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation Update (if none of the other choices apply)

## Checklist

_Put an `x` in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code._

- [x] Lint and unit tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have added the necessary documentation (if appropriate)
- [x] Any dependent changes have been merged and published in downstream
modules
- [x] Link to issue to close it